Use FXGL's internal clock in FXGL mode.
r4230
2017-11-13

# NEW 
1. In FXGL mode, switch from using ClockThreadTask in MasterClock 
   to FXGL's internal clock 
2. Add default_tps (Ticks per sec) in MasterClock.
3. Add onUpdate() in MarsProjectFXGL, Simulation, MasterClock.
4. Add getFPS() in MasterClock. Called by tps in MainScene.

# CHANGE 
1. Replace the use of Thread with Executor thread for saving sim 
   - Add saveExecutor and use it in saveOnExit(), saveSimulation(), 
     saveSimulationProcess() in MainScene.
2. Modify fireClockPulse() and statusUpdate() in MasterClock.
3. Reduce base-time-between-updates from 25 Hz to 20 Hz in 
   simulations.xml.
4. Reduce TIME_LIMIT in UpTimer from 3000 ms to 150 ms

+++ mars-sim-core/src/main/java/org/mars_sim/msp/core/time/EarthClock.java
@@ -321,10 +321,10 @@ implements Serializable {
 	 * Adds time to the calendar
 	 * @param ms milliseconds added to the calendar
 	 */
-	public void addTime(long ms) {
+	public void addTime(double ms) {
 		gregCal.add(Calendar.MILLISECOND, (int) ms);
 	}
-
+	
 	/**
 	 * Displays the string version of the clock.
 	 * @return time stamp string.
